gnn_vertex_finder/requirements.txt
comet_ml==3.31.3
numpy==1.22.3
pytorch_lightning==1.6.4
PyYAML==6.0
torch==1.11.0
torch_geometric==2.0.4
uproot==4.2.3
